+++ src/DDD/DDD/Commands/ICommandHandler.cs
@@ -1,6 +1,6 @@
 using MediatR;
 
-namespace GoldenEye.Backend.Core.DDD.Commands
+namespace GoldenEye.DDD.Commands
 {
     public interface ICommandHandler<in T>: IRequestHandler<T>
         where T : ICommand