+++ src/DDD/DDD/Events/EventBus.cs
@@ -2,7 +2,7 @@ using System.Threading;
 using System.Threading.Tasks;
 using MediatR;
 
-namespace GoldenEye.Backend.Core.DDD.Events
+namespace GoldenEye.DDD.Events
 {
     public class EventBus: IEventBus
     {