+++ src/DDD/DDD/Queries/IListQuery.cs
@@ -1,6 +1,6 @@
 using System.Collections.Generic;
 
-namespace GoldenEye.Backend.Core.DDD.Queries
+namespace GoldenEye.DDD.Queries
 {
     public interface IListQuery<TResponse>: IQuery<IReadOnlyList<TResponse>>
     {