+++ src/DDD/DDD/Queries/IQueryHandler.cs
@@ -1,6 +1,6 @@
 using MediatR;
 
-namespace GoldenEye.Backend.Core.DDD.Queries
+namespace GoldenEye.DDD.Queries
 {
     public interface IQueryHandler<in TQuery, TResponse>: IRequestHandler<TQuery, TResponse>
         where TQuery : IQuery<TResponse>