+++ src/DDD/DDD.Tests/Changelog.md
@@ -13,7 +13,7 @@
 ## Changes
 
 * Updated version of `MediatR` to `5.0.1` **[MAJOR]**
-* Added [tests](Registration/EventHandlerRegistrationTests.cs) for [EventHandler registration](../Backend.Core.DDD/Registration/Registration.cs) **[MINOR]**
+* Added [tests](Registration/EventHandlerRegistrationTests.cs) for [EventHandler registration](../DDD/Registration/Registration.cs) **[MINOR]**
 
 
 # v3.0.0 (29.12.2017) [Pull Request](https://github.com/oskardudycz/GoldenEye/pull/44)
@@ -28,13 +28,13 @@
 
 ## Changes
 
-* Added [tests](Validation/ValidationPipelineTests.cs) for [EventStorePipeline](../Backend.Core.DDD/Validation/ValidationPipeline.cs) **[MAJOR]**
+* Added [tests](Validation/ValidationPipelineTests.cs) for [EventStorePipeline](../DDD/Validation/ValidationPipeline.cs) **[MAJOR]**
 
 # v2.0.0
 
 ## Changes
 
-* Aded [tests](Events/Store/EventStorePipelineTests.cs) for [EventStorePipeline](../Backend.Core.DDD/Events/Store/EventStorePipeline.cs) **[MAJOR]**
+* Aded [tests](Events/Store/EventStorePipelineTests.cs) for [EventStorePipeline](../DDD/Events/Store/EventStorePipeline.cs) **[MAJOR]**
 
 # v1.0.0